Remove unneccessary conversions
I decided it would be like this after a bit of thought. Since I'm not
listing all of each objects attributes, the only documentation for the
types is on reddit's page. So the types stay close to what was described
and the user can do the conversion themselves (it's not that hard)