Unmarshaller should skip responses with no kind
Trying to run `flair_listing`, I ran into an error in Unmarshaller. Basically, it's trying to unmarshal something that's a hash, but doesn't have a `:kind` key in that hash.

I fixed it for my case by adding a check to see if response[:kind] even exists. If it's not there, there's no purpose behind trying to create an object.