Drop support for Ruby versions: 2.1, 2.2, 2.3, 2.4, 2.5
Ruby 2.1.0 was released: 2013-12-25
The first commit in this project was: 2017-01-14

The closest Ruby version I could find to early 2017 was: Ruby 2.4.0 that
was released in 2016-12-25.

That means you could technically run this on 5 different Ruby majors:
2.1, 2.2, 2.3, 2.4 and 2.5

I decided to drop these and move the required Ruby version to 2.6.
Making it possible to run this gem on the major versions: 2.6, 2.7, 3.0, 3.1